Update template.html
Removed unneeded ChromeFrame support - this has long since been discontinued by Google and is no longer available to download. (This PR is for Issue #24)

Added IE-edge X-UA-Compatible tag. This is recommended to prevent IE running in old-IE mode.
REF: https://www.modern.ie/en-gb/performance/how-to-use-x-ua-compatible

I've tested the above changes locally and live in production (and made a presentation using it at the weekend :-) )